+++ settings/ScanCheckoutSettings.js
@@ -24,7 +24,7 @@ class ScanCheckoutSettings extends React.Component {
     userIdentifierPref: {
       type: 'okapi',
       records: 'configs',
-      path: 'configurations/entries?query=(module=SCAN and config_name=pref_patron_identifier)',
+      path: 'configurations/entries?query=(module=SCAN and configName=pref_patron_identifier)',
       POST: {
         path: 'configurations/entries',
       },
@@ -52,7 +52,7 @@ class ScanCheckoutSettings extends React.Component {
       this.props.mutator.userIdentifierPref.POST(
         {
           module: 'SCAN',
-          config_name: 'pref_patron_identifier',
+          configName: 'pref_patron_identifier',
           value: e.target.value,
         },
       );