+++ settings/index.js
@@ -1,5 +1,3 @@
-// We have to remove node_modules/react to avoid having multiple copies loaded.
-// eslint-disable-next-line import/no-unresolved
 import React from 'react';
 import Settings from '@folio/stripes-components/lib/Settings';
 
@@ -19,4 +17,4 @@ const pages = [
   },
 ];
 
-export default props => <Settings {...props} pages={pages} />;
+export default props => <Settings {...props} pages={pages} paneTitle="Scan" />;